requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9
djangorestframework>=3.15
djangorestframework-simplejwt>=2.1
drf-spectacular>=1.1
django-cors-headers>=3.15
